+++ pkg/apis/workflow/v1alpha1/zz_generated.deepcopy.go
@@ -791,6 +791,11 @@ func (in *WorkflowSpec) DeepCopyInto(out *WorkflowSpec) {
 			(*in).DeepCopyInto(*out)
 		}
 	}
+	if in.ImagePullSecrets != nil {
+		in, out := &in.ImagePullSecrets, &out.ImagePullSecrets
+		*out = make([]v1.LocalObjectReference, len(*in))
+		copy(*out, *in)
+	}
 	return
 }