+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0-beta2 (Unreleased)
 + Add ability to specify affinity rules at both the workflow and template level
++ Add ability to specify imagePullSecrets in the workflow.spec
 - Fix issue preventing the referencing of artifacts in a container with retries
 - Fix issue preventing the use of volumes in a sidecar